+++ ft_printf.h
@@ -18,7 +18,7 @@
 # define HEXACAPS "0123456789ABCDEF"
 # define HEXAMIN "0123456789abcdef"
 
-int		ft_printhexa(unsigned long int hexa, char type);
+int		ft_printhexa(unsigned int hexa, char type);
 int		ft_printf_format(char format, va_list arg);
 int		ft_printf(const char *format, ...);
 char	*ft_itoa_unsigned(unsigned int nb);
